refactored trees, linked list and lifo fifo structures and their usage
