+++ basic_data_structures/linked_list/__init__.py
